Updated data layer tests to match code changes and moved to DataTests folder.
